Removed Amiga specific types and unnecessary includes.

+++ libmdouble/fixdfsi.c
@@ -1,15 +1,12 @@
-/*----------------*/
-/* libmathdoubbas */
-/*----------------*/
-
-#include <exec/types.h>
-#include <exec/libraries.h>
+/*------------*/
+/* libmdouble */
+/*------------*/
 
 extern struct Library *MathIeeeDoubBasBase;
 
-__stdargs LONG __fixdfsi(DOUBLE x)
+__stdargs long __fixdfsi(long x)
 {
-	register LONG r asm("d0");
+	register long r asm("d0");
 
 	asm("move.l 8(sp),d0; move.l 12(sp),d1; movea.l %1,a6; jsr -30(a6);" :
 		"=r"(r) : "m"(MathIeeeDoubBasBase) :
